# openapi-facade.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Custom GPT Facade API",
    "version": "2.1.0",
    "description": "BFF (Backend for Frontend) for Custom GPT actions. Provides high-level macros optimized for LLM consumption + unified RPC interface. Timezone: Europe/Prague. All relative time windows (today, lastHour, etc.) are resolved server-side."
  },
  "servers": [
    { "url": "https://api.example.com" }
  ],
...
  "paths": {
...
    "/rpc/mail": {
      "post": {
        "tags": ["rpc"],
        "operationId": "mailRpc",
        "summary": "Mail RPC: search, read, send, reply, modify",
        "description": "Unified mail operations. Op: search (Gmail query), preview (metadata only), read (full body), send (new + draft), reply, modify (mark/star/label), attachmentPreview (text/table), labels (list/modify). Always includes attachment URLs (expire 1h). Blocked attachments have blocked:true, no URL.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MailRpcRequest" },
              "examples": {
                "send-draft": {
                  "summary": "Send existing draft by ID",
                  "value": {
                    "op": "send",
                    "params": {
                      "draftId": "DRAFT_123"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operation result + optional warnings",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/MailRpcResult" } }
            }
          },
          "400": { "$ref": "#/components/responses/ClientError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimit429" },
          "451": { "$ref": "#/components/responses/AttachmentBlocked" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/rpc/calendar": {
      "post": {
        "tags": ["rpc"],
        "operationId": "calendarRpc",
        "summary": "Calendar RPC: CRUD + conflict checking",
        "description": "Low-level calendar ops. Op: list, get, create, update, delete, checkConflicts. Use RPC when macros don't fit. checkConflicts validates time windows before scheduling. Returns detailed conflict info (what event is blocking).",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CalendarRpcRequest" } } }
        },
        "responses": {
          "200": { "description": "Calendar operation result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CalendarRpcResult" } } } },
          "400": { "$ref": "#/components/responses/ClientError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/ConflictWithAlternatives" },
          "429": { "$ref": "#/components/responses/RateLimit429" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/rpc/contacts": {
      "post": {
        "tags": ["rpc"],
        "operationId": "contactsRpc",
        "summary": "Contacts RPC: search, CRUD, deduplication",
        "description": "Low-level contact ops. Op: list, search, add, update, delete, dedupe, bulkUpsert. Dedupe scans for name/email/phone similarity (0-1 score). Use when safeAdd macro insufficient. Max 100 items per bulk request.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ContactsRpcRequest" } } } },
        "responses": {
          "200": { "description": "Contacts operation result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ContactsRpcResult" } } } },
          "400": { "$ref": "#/components/responses/ClientError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimit429" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/rpc/tasks": {
      "post": {
        "tags": ["rpc"],
        "operationId": "tasksRpc",
        "summary": "Tasks RPC: CRUD + status toggle",
        "description": "Low-level task ops. Op: list, get, create, update, delete, complete, reopen. Complete/reopen are status shortcuts. Use RPC for bulk operations or fine-grained control beyond /macros/tasks/overview.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TasksRpcRequest" } } } },
        "responses": {
          "200": { "description": "Tasks operation result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TasksRpcResult" } } } },
          "400": { "$ref": "#/components/responses/ClientError" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimit429" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "RelativeTime": {
        "type": "string",
        "description": "Relative time window resolved server-side (Europe/Prague TZ). Use instead of manual datetimes for better UX and DST handling.",
        "enum": ["lastHour", "last3h", "last24h", "today", "yesterday", "thisWeek", "last7d"]
      },
      "TimeRange": {
        "oneOf": [
          {
            "type": "object",
            "description": "Relative time window (preferred for GPT queries)",
            "properties": { "relative": { "$ref": "#/components/schemas/RelativeTime" } },
            "required": ["relative"],
            "additionalProperties": false
          },
          {
            "type": "object",
            "description": "Absolute time window (ISO 8601, UTC)",
            "properties": {
              "start": { "type": "string", "format": "date-time", "description": "Inclusive start (ISO 8601)." },
              "end": { "type": "string", "format": "date-time", "description": "Inclusive end (ISO 8601)." }
            },
            "required": ["start", "end"],
            "additionalProperties": false
          }
        ]
      },
...
      "MailRpcRequest": {
        "type": "object",
        "required": ["op", "params"],
        "properties": {
          "op": { "type": "string", "enum": ["search", "preview", "read", "send", "reply", "modify", "attachmentPreview", "labels"], "description": "Operation: search (query), preview (metadata only), read (full), send (new/draft), reply, modify (mark/star/label), attachmentPreview, labels (list/update)." },
          "params": { "oneOf": [ { "$ref": "#/components/schemas/MailSearchParams" }, { "$ref": "#/components/schemas/MailPreviewParams" }, { "$ref": "#/components/schemas/MailReadParams" }, { "$ref": "#/components/schemas/MailSendParams" }, { "$ref": "#/components/schemas/MailReplyParams" }, { "$ref": "#/components/schemas/MailModifyParams" }, { "$ref": "#/components/schemas/MailAttachmentPreviewParams" }, { "$ref": "#/components/schemas/MailLabelsParams" } ] }
        }
      },
      "MailSearchParams": {
        "type": "object",
        "description": "Gmail-like search query (from:, to:, subject:, has:attachment, label:, after:date, etc.).",
        "properties": {
          "query": { "type": "string", "maxLength": 255 },
          "timeRange": { "$ref": "#/components/schemas/TimeRange" },
          "maxResults": { "type": "integer", "minimum": 1, "maximum": 100, "default": 50 }
        },
        "required": ["query"]
      },
      "MailPreviewParams": {
        "type": "object",
        "description": "Get metadata-only for specified message IDs (fastest).",
        "properties": { "ids": { "type": "array", "items": { "type": "string" }, "minItems": 1, "maxItems": 50 } },
        "required": ["ids"]
      },
      "MailReadParams": {
        "type": "object",
        "description": "Provide ids[] or searchQuery (not both). Returns single item or batch.",
        "properties": {
          "ids": { "type": "array", "items": { "type": "string" }, "minItems": 1, "maxItems": 50 },
          "searchQuery": { "type": "string", "maxLength": 255 },
          "format": { "type": "string", "enum": ["snippet", "minimal", "metadata", "full"], "default": "minimal" }
        }
      },
      "MailSendParams": {
        "oneOf": [
          {
            "type": "object",
            "description": "Send new email.",
            "required": ["to", "subject", "bodyText"],
            "properties": {
              "to": { "type": "array", "minItems": 1, "items": { "type": "string", "format": "email" }, "maxItems": 50 },
              "cc": { "type": "array", "items": { "type": "string", "format": "email" }, "maxItems": 50 },
              "bcc": { "type": "array", "items": { "type": "string", "format": "email" }, "maxItems": 50 },
              "subject": { "type": "string", "maxLength": 255 },
              "bodyText": { "type": "string", "maxLength": 20000 },
              "bodyHtml": { "type": ["string", "null"], "maxLength": 30000 },
              "toSelf": { "type": "boolean", "default": false },
              "confirmSelfSend": { "type": "boolean", "default": false, "description": "Required=true if toSelf=true." }
            },
            "additionalProperties": false
          },
          {
            "type": "object",
            "description": "Send existing Gmail draft.",
            "required": ["draftId"],
            "properties": {
              "draftId": { "type": "string", "description": "Gmail draft ID to send." }
            },
            "additionalProperties": false
          }
        ]
      },
      "MailReplyParams": {
        "type": "object",
        "required": ["messageId", "bodyText"],
        "properties": {
          "messageId": { "type": "string" },
          "bodyText": { "type": "string", "maxLength": 20000 },
          "bodyHtml": { "type": ["string", "null"], "maxLength": 30000 },
          "includeQuote": { "type": "boolean", "default": true }
        }
      },
      "MailModifyParams": {
        "type": "object",
        "required": ["ids", "actions"],
        "properties": {
          "ids": { "type": "array", "items": { "type": "string" }, "minItems": 1, "maxItems": 100 },
          "actions": {
            "type": "object",
            "properties": {
              "markRead": { "type": "boolean" },
              "star": { "type": "boolean" },
              "addLabels": { "type": "array", "items": { "type": "string" }, "maxItems": 20 },
              "removeLabels": { "type": "array", "items": { "type": "string" }, "maxItems": 20 }
            }
          }
        }
      },
      "MailAttachmentPreviewParams": {
        "type": "object",
        "required": ["messageId", "attachmentId", "mode"],
        "properties": {
          "messageId": { "type": "string" },
          "attachmentId": { "type": "string" },
          "mode": { "type": "string", "enum": ["text", "table"], "description": "text: PDF/TXT/HTML (max 256KB). table: CSV/XLSX (max 200 rows)." },
          "maxSize": { "type": "integer", "minimum": 1, "maximum": 256, "default": 256, "description": "Max KB (text) or max rows (table)." },
          "delimiter": { "type": "string", "enum": ["auto", "comma", "semicolon", "tab"], "default": "auto", "description": "CSV delimiter (CSV only)." }
        }
      },
      "MailLabelsParams": {
        "type": "object",
        "properties": {
          "list": { "type": "boolean", "description": "If true: return all labels. If false: modify message labels." },
          "modify": {
            "type": "object",
            "properties": {
              "messageId": { "type": "string", "description": "Message ID (or use ids for batch)." },
              "ids": { "type": "array", "items": { "type": "string" }, "maxItems": 100 },
              "add": { "type": "array", "items": { "type": "string" }, "maxItems": 20 },
              "remove": { "type": "array", "items": { "type": "string" }, "maxItems": 20 }
            }
          }
        }
      },
      "MailRpcResult": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean" },
          "data": { "type": ["object", "array", "null"] },
          "securityWarnings": { "type": "array", "items": { "type": "string" }, "description": "Warnings about blocked attachments, suspicious content, etc." }
        },
        "required": ["ok"]
      },
      "CalendarRpcRequest": {
        "type": "object",
        "required": ["op", "params"],
        "properties": {
          "op": { "type": "string", "enum": ["list", "get", "create", "update", "delete", "checkConflicts"] },
          "params": { "type": "object" }
        }
      },
      "CalendarRpcResult": { "type": "object", "properties": { "ok": { "type": "boolean" }, "data": { "type": ["object", "array", "null"] } }, "required": ["ok"] },
      "ContactsRpcRequest": {
        "type": "object",
        "required": ["op", "params"],
        "properties": {
          "op": { "type": "string", "enum": ["list", "search", "add", "update", "delete", "dedupe", "bulkUpsert"] },
          "params": { "type": "object" }
        }
      },
      "ContactsRpcResult": { "type": "object", "properties": { "ok": { "type": "boolean" }, "data": { "type": ["object", "array", "null"] } }, "required": ["ok"] },
      "TasksRpcRequest": {
        "type": "object",
        "required": ["op", "params"],
        "properties": {
          "op": { "type": "string", "enum": ["list", "get", "create", "update", "delete", "complete", "reopen"] },
          "params": { "type": "object" }
        }
      },
      "TasksRpcResult": { "type": "object", "properties": { "ok": { "type": "boolean" }, "data": { "type": ["object", "array", "null"] } }, "required": ["ok"] }
    },
    "responses": {
      "ClientError": {
        "description": "4xx client error (invalid request, missing params, enum violation, etc.)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string" },
                "message": { "type": "string", "maxLength": 300 },
                "code": { "type": "string", "enum": ["INVALID_PARAM", "BAD_REQUEST", "CONFIRM_SELF_SEND_REQUIRED", "CONFIRMATION_EXPIRED"] },
                "hint": { "type": ["string", "null"], "maxLength": 200 }
              },
              "required": ["error", "message"]
            }
          }
        }
      },
      "Unauthorized": {
        "description": "401 - Not authenticated or token expired",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string" },
                "message": { "type": "string" }
              },
              "required": ["error", "message"]
            }
          }
        }
      },
...
      "ConflictWithAlternatives": {
        "description": "409 - Scheduling conflict with alternative suggestions",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string" },
                "message": { "type": "string" },
                "alternatives": { "type": "array", "items": { "type": "object", "properties": { "start": { "type": "string", "format": "date-time" }, "end": { "type": "string", "format": "date-time" } }, "required": ["start", "end"] }, "maxItems": 5 }
              },
              "required": ["error", "message"]
            }
          }
        }
      },
      "AttachmentBlocked": {
        "description": "451 - Attachment URL withheld due to security policy",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string" },
                "message": { "type": "string" },
                "code": { "type": "string", "const": "ATTACHMENT_BLOCKED" }
              },
              "required": ["error", "message", "code"]
            }
          }
        }
      },
      "ServerError": {
        "description": "5xx server error (temporary issue, please retry)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string" },
                "message": { "type": "string" }
              },
              "required": ["error", "message"]
            }
          }
        }
      },
      "RateLimit429": {
        "description": "Too Many Requests - rate limit exceeded",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": { "type": "string", "example": "rate_limit_exceeded" },
                "retryAfterSeconds": { "type": "integer", "minimum": 1 }
              },
              "required": ["error"]
            }
          }
        },
        "headers": {
          "Retry-After": {
            "description": "Seconds to wait before retrying",
            "schema": { "type": "integer", "minimum": 1 }
          }
        }
      }
    }
  }
}